chore(types): centralize models, move TimelineEntry to src/types/timeline.ts, update imports to '@/types'
